Add pure Ruby fallback from Sonic Pi
Given that we spent so much time optimizing the pure Ruby version, it
seems sensible to expose it in some form here. This also helps in the
case of `LoadError`s where the gem has been compiled and distributed to
incompatible platforms. Can be forced via env var as described in
README.