Add decoding of OSC bundles
The new function `decode_bundle`, returns `(timestamp, msgs)`, where msgs
contains all the messages in the bundle. In case that it is called on
something that is not really a bundle it will return `nil`.  There is also
a function `is_bundle?` to check if the packet is really a bundle.