Change API to be more consistent with bundles
Provide a decode method that will work on bundles and discrete messages,
returning a list of [timestamp, [messages in bundle]], where each
message is [path, arguments]. This is necessary because of the atomicity
requirements of OSC bundles.

The representation is a list, not a tree, because the atomicity is not
to be preserved for enclosing bundles, only at the leaf level. The list
makes it easier to process for the application.

A simpler and faster decode_no_bundles is provided, in case the
a[plication knows that there will be no bundles at all.