Use MSYS2 toolchain to build on windows, not VS/nmake
This abandons using nmake and cl (the tools installed as part of Visual
Studio) and goes back to using `-rdevkit` instead. Devkit is part of the
RubyInstaller package that comes with the AppVeyor image.

It's better to build C extensions with the same toolchain that was used
to build the Ruby version to avoid conflicts.